Accept big (above 2g) ints in task start and end range offsets (#359)

+++ server/collab/models.py
@@ -126,8 +126,8 @@ class Task(models.Model):
   source_file_version = models.ForeignKey(FileVersion,
                                           related_name='source_tasks')
   # TODO: make sure start > end
-  source_start = models.PositiveIntegerField(null=True)
-  source_end = models.PositiveIntegerField(null=True)
+  source_start = models.BigIntegerField(null=True)
+  source_end = models.BigIntegerField(null=True)
   # TODO: make sure only at least one of target_file/target_project is null
   target_file = models.ForeignKey(File, null=True)
   target_project = models.ForeignKey(Project, null=True)
